+++ expectation.messagetype.go
@@ -134,9 +134,9 @@ func (p *messageTypePredicate) Ok() bool {
 func (p *messageTypePredicate) Done() {
 }
 
-func (p *messageTypePredicate) Report(treeOk, isInverted bool) *Report {
+func (p *messageTypePredicate) Report(ctx ReportGenerationContext) *Report {
 	rep := &Report{
-		TreeOk: treeOk,
+		TreeOk: ctx.TreeOk,
 		Ok:     p.ok,
 		Criteria: inflect.Sprintf(
 			p.expectedRole,
@@ -145,7 +145,7 @@ func (p *messageTypePredicate) Report(treeOk, isInverted bool) *Report {
 		),
 	}
 
-	if treeOk || p.ok || isInverted {
+	if p.ok || ctx.TreeOk || ctx.IsInverted {
 		return rep
 	}