+++ CHANGELOG.md
@@ -10,6 +10,18 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+### Added
+
+- Added `ReportGenerationContext` to hold information required when generating
+  test reports.
+
+### Changed
+
+- **[BC]** Changed `Predicate.Report()` to accept a `ReportGenerationContext`
+  instead of individual arguments.
+
 ## [0.16.0] - 2024-08-17
 
 ### Changed